Add files via upload
This Task required making the chart easily digestible. Making it clear to tell which region salesmen are from, and who was the top salesmen in each region. 
To make this possible the region and representatives were added as Dimensions. Additionally, a calculated field was created to measure the total sales in dollar amount.